refactoring, beginning implementation of botmanager-api
