+++ app/index.ts
@@ -5,7 +5,7 @@ import player_router from './routes/player-routes.ts';
 import db_service from './services/db-service.ts';
 
 const app = express();
-/* const port = 8080;
+const port = 8080;
 
 app.use(express.json());
 
@@ -13,19 +13,18 @@ app.get('/', (req: any, res:any) => {
     res.json({'message': 'ok'});
 })
 
-app.use('/player', player_router); */
+app.use('/player', player_router);
 
-async function startServer() {
+async function startServer(port: number) {
     await db_service.init();
     await db_service.createTables();
-    /* return new Promise<void>((resolve) => {
-        app.listen(port, ()  => {
+    return new Promise<void>((resolve) => {
+        app.listen(port, () => {
             console.log(`App listening on http://localhost:${port}`);
-            return resolve();
+            resolve();
         });
-    }); */
+    });
 }
 
-startServer().then(
-    async() => await bot_manager()
-)
+await startServer(port);
+await bot_manager();
