+++ app/services/api/botmanagerapi-service.ts
@@ -1,0 +1,6 @@
+
+export class PlayerService {
+    constructor() {
+    }
+    // NOT IMPLEMENTED
+}
