+++ app/services/api/playerapi-service.ts
@@ -1,5 +1,5 @@
-import { DBService} from './db-service.ts';
-import { emptyOrSingleRow } from '../helpers/db-query-helper.ts'
+import { DBService} from '../db-service.ts';
+import { emptyOrSingleRow } from '../../helpers/db-query-helper.ts'
 
 export class PlayerService {
     private db_service: DBService;
