+++ app/services/log-service.ts
@@ -1,5 +1,5 @@
 import puppeteer from 'puppeteer';
-import type { Response } from '../utils/error-handling-utils.ts';
+import type { Response } from '../utils/errorhandling-utils.ts';
 import { Data, Log } from '../interfaces/log-processing-interfaces.ts';
 
 export class LogService {
