+++ app/utils/datastructures.ts
